requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
enum-compat==0.0.3
idna==3.7
intuit-oauth==1.2.5
numpy==2.0.1
oauthlib==3.2.2
pandas==2.2.2
psycopg2==2.9.9
pyasn1==0.6.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
XlsxWriter==3.2.0
